refatorando TipoPagamento

+++ #Banco de Dados/03 - CREATE TABLE - TipoPagamento.sql
@@ -3,7 +3,7 @@ USE [ControleDespesas]
 IF NOT EXISTS (SELECT 1 FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_TYPE='BASE TABLE' AND TABLE_NAME='TipoPagamento') 
 BEGIN
     CREATE TABLE [dbo].[TipoPagamento](
-	[Id] [int] IDENTITY(1,1) NOT NULL,
+	[Id] [bigint] IDENTITY(1,1) NOT NULL,
 	[Descricao] [nvarchar](250) NOT NULL,
     CONSTRAINT [PK_TipoPagamento] PRIMARY KEY CLUSTERED 
 (
